+++ SteamTogether.Bot/Migrations/ApplicationDbContextModelSnapshot.cs
@@ -21,7 +21,7 @@ namespace SteamTogether.Bot.Migrations
                     b.Property<string>("PlayerId")
                         .HasColumnType("TEXT");
 
-                    b.Property<int>("ChatId")
+                    b.Property<long>("ChatId")
                         .HasColumnType("INTEGER");
 
                     b.HasKey("PlayerId", "ChatId");
@@ -34,12 +34,12 @@ namespace SteamTogether.Bot.Migrations
                         new
                         {
                             PlayerId = "76561198068819558",
-                            ChatId = 1
+                            ChatId = 1L
                         },
                         new
                         {
                             PlayerId = "zebradil",
-                            ChatId = 1
+                            ChatId = 1L
                         });
                 });
 
@@ -68,7 +68,7 @@ namespace SteamTogether.Bot.Migrations
 
             modelBuilder.Entity("SteamTogether.Bot.Models.TelegramChat", b =>
                 {
-                    b.Property<int>("ChatId")
+                    b.Property<long>("ChatId")
                         .ValueGeneratedOnAdd()
                         .HasColumnType("INTEGER");
 
@@ -79,7 +79,7 @@ namespace SteamTogether.Bot.Migrations
                     b.HasData(
                         new
                         {
-                            ChatId = 1
+                            ChatId = 1L
                         });
                 });
